split streams file into several files
