+++ source/views/streaming/webcams/types.js
@@ -1,0 +1,11 @@
+// @flow
+
+export type Webcam = {
+  streamUrl: string,
+  pageUrl: string,
+  name: string,
+  thumbnail: string,
+  thumbnailUrl?: string,
+  textColor: string,
+  accentColor: [number, number, number],
+}
